+++ src/LifeOS.Common/ViewModels/DayPlan/DayPlanViewModel.DropTarget.cs
@@ -1,20 +1,12 @@
 ﻿using GongSolutions.Wpf.DragDrop;
-using LifeOS.Common.Models;
-using LifeOS.Common.Sys.ComponentModel;
+using LifeOS.Common.ViewModels.DayActivityVM;
 using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.ComponentModel;
-using System.Linq;
 using System.Text;
-using System.Text.Json.Serialization;
-using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Input;
 
-namespace LifeOS.Common.ViewModels
+namespace LifeOS.Common.ViewModels.DayPlanVM
 {
-
   public partial class DayPlanViewModel : IDropTarget
   {
 
@@ -35,24 +27,6 @@ namespace LifeOS.Common.ViewModels
       DayActivityViewModel sourceItem = dropInfo.Data as DayActivityViewModel;
       DayPlanViewModel targetItem = dropInfo.TargetItem as DayPlanViewModel;
       targetItem.Activities.Add(sourceItem);
-    }
-
-  }
-
-  public partial class DayPlanViewModel : BaseViewModel
-  {
-
-    public DateTime Day { get; set; }
-    public bool CanAcceptChildren { get; set; } = true;
-    public ObservableCollection<DayActivityViewModel> Activities { get; set; }
-
-    public DayPlanViewModel(DayPlan plan)
-    {
-
-      this.Day = plan.Day;
-      this.Activities = new ObservableCollection<DayActivityViewModel>(
-        plan.Activities.Select(x => new DayActivityViewModel(x)));
-
     }
 
   }
