+++ src/LifeOS.WPF/Views/Windows/MainWindow.xaml.cs
@@ -40,6 +40,8 @@ namespace LifeOS.WPF.Views.Windows
   using System.Windows.Media;
   using LifeOS.Common.Models;
   using LifeOS.Common.ViewModels;
+  using LifeOS.Common.ViewModels.DayActivityVM;
+  using LifeOS.Common.ViewModels.DayPlanVM;
   using MahApps.Metro.Controls;
   using NodaTime;
